buffer: improve buff.toString(encoding, start, end)
When buff.toString() without parameter will call
buff.utf8Slice(0, this.length) directly. It skips the start/end
check, even thougth the buffer length is zero.

The utf8Slice is a C++ binding method, when the buffer length is
zero, the case is unexpected.

And, the slowToString.apply(this, arguments) is slowly also,
change to slowToString.call(this, encoding, start, end).